pyproject.toml
[tool.poetry]
name = "poc-llm"
version = "0.1.0"
description = ""
authors = ["Mikhail Ozernov <ozzygogogo@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "llm_poc", from = "src" }]

[tool.poetry.dependencies]
python = "3.11.7"
google-genai = "^1.38.0"
dotenv = "^0.9.9"
scikit-learn = "^1.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
llm-poc = "llm_poc.__main__:main"

